[FEAT] make Remote::Net a role
Remote::Net offers functions to send state messages to MCP. The using
module has three possible ways of using it:
* extending
* role
* Module->new();

Extending is wrong in object hierarchy because an Installer object is
not a Remote::Net object somehow. Using new means we can not have
additional state sender functions in the same object (would need a
$remote_net and a $other_net object). Creating an Installer::Net object
would have the same issues.

Thus making it a role is the right choice.